+++ content/ua/1-5-Decomposition.md
