+++ content/ua/1-7-Paradigms.md
