+++ content/ua/2-2-Types.md
