+++ content/ua/2-6-Function.md
