+++ content/ua/2-7-Closure.md
