+++ content/ua/2-9-Monomorphic.md
