+++ content/ua/3-5-Deque.md
