+++ content/ua/4-1-Tech-stack.md
