+++ content/ua/4-6-Partial.md
